Move build-update-utils to plugin
