+++ unified-prototype/unified-plugin/build-logic/build.gradle.kts
@@ -5,3 +5,12 @@ plugins {
 dependencies {
     implementation("com.gradle.publish:plugin-publish-plugin:1.2.1")
 }
+
+gradlePlugin {
+    plugins {
+        create("build-logic.build-update-utils") {
+            id = "build-logic.build-update-utils"
+            implementationClass = "buildlogic.BuildUpdateUtilsPlugin"
+        }
+    }
+}
